issue #4 - optional tableIdColumns configuration
This allows for an optional configuration map of table name to column name to
use for the id column. If a table is not specified in the map it will default to
the original value of 'id'.

Also added improved validation to check all tables for the right id column.